revert: move base input components back to components folder
