+++ client/src/js/app/components/base-input-checkbox.vue
